megamarket/requirements.txt
Django==2.2.16
django-simple-history==3.1.1
djangorestframework==3.12.4
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
gunicorn==20.0.4
isort==5.10.1
mccabe==0.6.1
pep8-naming==0.13.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
testfixtures==6.18.5
